[fix] slack sending incorrect message with travis
Depending on the script we run we may not be able to generate the log file which contain the error and warning messages. In this case we need to rely on the CI tool to tell us at least that it failed. The build is currently fail so we want to make sure that slack is able to send the correct message.