[feat] allow isRelease and isReleasePullRequest methods to be async
This is needed so that we can treat different builds as releases. Now that we want to support node 8 and 10 (for when we upgrade angular). Currently the fp repo is failing because one of the node builds released and the other one was denied. Now we can specify the release tied to only one build.